Lowering of transforming-for
